+++ lib/zendesk_messaging.dart
@@ -16,38 +16,39 @@ enum ZendeskMessagingMessageType {
 
 /// Used by ZendeskMessaging to attach custom async observers
 class ZendeskMessagingObserver {
+  ZendeskMessagingObserver(this.removeOnCall, this.func);
   final bool removeOnCall;
   final Function(Map? args) func;
-
-  ZendeskMessagingObserver(this.removeOnCall, this.func);
 }
 
 class ZendeskLoginResponse {
+  ZendeskLoginResponse(this.id, this.externalId);
   final String? id;
   final String? externalId;
-
-  ZendeskLoginResponse(this.id, this.externalId);
 }
 
 class ZendeskMessaging {
   static const MethodChannel _channel = MethodChannel('zendesk_messaging');
   static const channelMethodToMessageType = {
-    "initialize_success": ZendeskMessagingMessageType.initializeSuccess,
-    "initialize_failure": ZendeskMessagingMessageType.initializeFailure,
-    "login_success": ZendeskMessagingMessageType.loginSuccess,
-    "login_failure": ZendeskMessagingMessageType.loginFailure,
-    "logout_success": ZendeskMessagingMessageType.logoutSuccess,
-    "logout_failure": ZendeskMessagingMessageType.logoutFailure,
+    'initialize_success': ZendeskMessagingMessageType.initializeSuccess,
+    'initialize_failure': ZendeskMessagingMessageType.initializeFailure,
+    'login_success': ZendeskMessagingMessageType.loginSuccess,
+    'login_failure': ZendeskMessagingMessageType.loginFailure,
+    'logout_success': ZendeskMessagingMessageType.logoutSuccess,
+    'logout_failure': ZendeskMessagingMessageType.logoutFailure,
   };
 
   /// Global handler, all channel method calls will trigger this observer
   static Function(ZendeskMessagingMessageType type, Map? arguments)? _handler;
 
   /// Allow end-user to use local observer when calling some methods
-  static final Map<ZendeskMessagingMessageType, ZendeskMessagingObserver> _observers = {};
+  static final Map<ZendeskMessagingMessageType, ZendeskMessagingObserver>
+      _observers = {};
 
   /// Attach a global observer for incoming messages
-  static void setMessageHandler(Function(ZendeskMessagingMessageType type, Map? arguments)? handler) {
+  static void setMessageHandler(
+    Function(ZendeskMessagingMessageType type, Map? arguments)? handler,
+  ) {
     _handler = handler;
   }
 
@@ -55,14 +56,19 @@ class ZendeskMessaging {
   ///
   /// @param  androidChannelKey  The Android SDK key generated from Zendesk dashboard
   /// @param  iosChannelKey      The iOS SDK key generated from the Zendesk dashboard
-  static Future<void> initialize({required String androidChannelKey, required String iosChannelKey}) async {
+  static Future<void> initialize({
+    required String androidChannelKey,
+    required String iosChannelKey,
+  }) async {
     if (androidChannelKey.isEmpty || iosChannelKey.isEmpty) {
       debugPrint('ZendeskMessaging - initialize - keys can not be empty');
       return;
     }
 
     try {
-      _channel.setMethodCallHandler(_onMethodCall); // start observing channel messages
+      _channel.setMethodCallHandler(
+        _onMethodCall,
+      ); // start observing channel messages
       await _channel.invokeMethod('initialize', {
         'channelKey': Platform.isAndroid ? androidChannelKey : iosChannelKey,
       });
@@ -74,7 +80,7 @@ class ZendeskMessaging {
   }
 
   /// Invalidates the current instance of ZendeskMessaging.
-/// After calling this method you will have to call ZendeskMessaging.initialize again if you would like to use ZendeskMessaging.
+  /// After calling this method you will have to call ZendeskMessaging.initialize again if you would like to use ZendeskMessaging.
   static Future<void> invalidate() async {
     try {
       await _channel.invokeMethod('invalidate');
@@ -93,24 +99,23 @@ class ZendeskMessaging {
   }
 
   /// Add a list of tags to a support ticket
-  /// 
-  /// Conversation tags are not immediately associated with a conversation when this method is called. 
-  /// It will only be applied to a conversation when end users either start a new 
+  ///
+  /// Conversation tags are not immediately associated with a conversation when this method is called.
+  /// It will only be applied to a conversation when end users either start a new
   /// conversation or send a new message in an existing conversation.
-  /// 
+  ///
   /// For example, to apply "promo_code" and "discount" tags to a conversation about an order, then you would call:
   /// `ZendeskMessaging.setConversationTags(["promo_code","discount"])`
   static Future<void> setConversationTags(List<String> tags) async {
     try {
-      await _channel.invokeMethod('setConversationTags',
-          {'tags': tags});
+      await _channel.invokeMethod('setConversationTags', {'tags': tags});
     } catch (e) {
       debugPrint('ZendeskMessaging - setConversationTags - Error: $e}');
     }
   }
 
-/// Remove all the tags on the current support ticket
-/// 
+  /// Remove all the tags on the current support ticket
+  ///
   static Future<void> clearConversationTags() async {
     try {
       await _channel.invokeMethod('clearConversationTags');
@@ -124,17 +129,31 @@ class ZendeskMessaging {
   /// @param  jwt       Required by the SDK - You must generate it from your backend
   /// @param  onSuccess Optional - If you need to be notified about the login success
   /// @param  onFailure Optional - If you need to be notified about the login failure
-  static Future<void> loginUserCallbacks(
-      {required String jwt, Function(String? id, String? externalId)? onSuccess, Function()? onFailure}) async {
+  static Future<void> loginUserCallbacks({
+    required String jwt,
+    Function(String? id, String? externalId)? onSuccess,
+    Function()? onFailure,
+  }) async {
     if (jwt.isEmpty) {
       debugPrint('ZendeskMessaging - loginUser - jwt can not be empty');
       return;
     }
 
     try {
-      _setObserver(ZendeskMessagingMessageType.loginSuccess,
-          onSuccess != null ? (Map? args) => onSuccess(args?["id"], args?["externalId"]) : null);
-      _setObserver(ZendeskMessagingMessageType.loginFailure, onFailure != null ? (Map? args) => onFailure() : null);
+      _setObserver(
+        ZendeskMessagingMessageType.loginSuccess,
+        onSuccess != null
+            ? (Map? args) {
+                final id = args?['id'] ?? '';
+                final externalId = args?['externalId'] ?? '';
+                onSuccess(id, externalId);
+              }
+            : null,
+      );
+      _setObserver(
+        ZendeskMessagingMessageType.loginFailure,
+        onFailure != null ? (Map? args) => onFailure() : null,
+      );
 
       await _channel.invokeMethod('loginUser', {'jwt': jwt});
     } catch (e) {
@@ -146,11 +165,13 @@ class ZendeskMessaging {
   ///
   /// @return   The zendesk userId
   static Future<ZendeskLoginResponse> loginUser({required String jwt}) async {
-    var completer = Completer<ZendeskLoginResponse>();
+    final completer = Completer<ZendeskLoginResponse>();
     await loginUserCallbacks(
       jwt: jwt,
-      onSuccess: (id, externalId) => completer.complete(ZendeskLoginResponse(id, externalId)),
-      onFailure: () => completer.completeError(Exception("Zendesk::loginUser failed")),
+      onSuccess: (id, externalId) =>
+          completer.complete(ZendeskLoginResponse(id, externalId)),
+      onFailure: () =>
+          completer.completeError(Exception('Zendesk::loginUser failed')),
     );
     return completer.future;
   }
@@ -159,10 +180,19 @@ class ZendeskMessaging {
   ///
   /// @param  onSuccess Optional - If you need to be notified about the logout success
   /// @param  onFailure Optional - If you need to be notified about the logout failure
-  static Future<void> logoutUserCallbacks({Function()? onSuccess, Function()? onFailure}) async {
+  static Future<void> logoutUserCallbacks({
+    Function()? onSuccess,
+    Function()? onFailure,
+  }) async {
     try {
-      _setObserver(ZendeskMessagingMessageType.logoutSuccess, onSuccess != null ? (Map? args) => onSuccess() : null);
-      _setObserver(ZendeskMessagingMessageType.logoutFailure, onFailure != null ? (Map? args) => onFailure() : null);
+      _setObserver(
+        ZendeskMessagingMessageType.logoutSuccess,
+        onSuccess != null ? (Map? args) => onSuccess() : null,
+      );
+      _setObserver(
+        ZendeskMessagingMessageType.logoutFailure,
+        onFailure != null ? (Map? args) => onFailure() : null,
+      );
 
       await _channel.invokeMethod('logoutUser');
     } catch (e) {
@@ -172,15 +202,16 @@ class ZendeskMessaging {
 
   /// Helper function to logout waiting for future to complete
   static Future<void> logoutUser() async {
-    var completer = Completer<void>();
+    final completer = Completer<void>();
     await logoutUserCallbacks(
-      onSuccess: () => completer.complete(),
-      onFailure: () => completer.completeError(Exception("Zendesk::logoutUser failed")),
+      onSuccess: completer.complete,
+      onFailure: () =>
+          completer.completeError(Exception('Zendesk::logoutUser failed')),
     );
     return completer.future;
   }
 
-  /// Retrieve uread messages count from the Zendesk SDK
+  /// Retrieve unread messages count from the Zendesk SDK
   static Future<int> getUnreadMessageCount() async {
     try {
       return await _channel.invokeMethod(
@@ -203,6 +234,7 @@ class ZendeskMessaging {
       return false;
     }
   }
+
   ///  Check if the user is already logged in
   static Future<bool> isLoggedIn() async {
     try {
@@ -221,14 +253,14 @@ class ZendeskMessaging {
       return;
     }
 
-    final ZendeskMessagingMessageType type = channelMethodToMessageType[call.method]!;
-    var globalHandler = _handler;
+    final type = channelMethodToMessageType[call.method]!;
+    final globalHandler = _handler;
     if (globalHandler != null) {
       globalHandler(type, call.arguments);
     }
 
     // call all observers too
-    final ZendeskMessagingObserver? observer = _observers[type];
+    final observer = _observers[type];
     if (observer != null) {
       observer.func(call.arguments);
       if (observer.removeOnCall) {
@@ -238,7 +270,11 @@ class ZendeskMessaging {
   }
 
   /// Add an observer for a specific type
-  static _setObserver(ZendeskMessagingMessageType type, Function(Map? args)? func, {bool removeOnCall = true}) {
+  static _setObserver(
+    ZendeskMessagingMessageType type,
+    Function(Map? args)? func, {
+    bool removeOnCall = true,
+  }) {
     if (func == null) {
       _observers.remove(type);
     } else {